dmel_codec/requirements.txt
einops==0.8.0
huggingface_hub==0.26.2
hydra-core==1.3.2
librosa==0.10.2.post1
lightning==2.4.0
lightning_utilities==0.11.9
matplotlib==3.10.0
numpy==2.2.0
omegaconf==2.3.0
pedalboard==0.9.16
rich==13.9.4
rootutils==1.0.7
scipy==1.14.1
soundfile==0.12.1
torchaudio==2.5.1
vector_quantize_pytorch==1.20.11
